Move DB helper function
